requirements.txt
dspy-ai>=2.0.0
pydantic>=2.0.0
requests>=2.25.0
numpy>=1.20.0
scikit-learn>=1.0.0
openai>=1.0.0
python-dotenv>=0.19.0 